# perf: improve encoding