--- a/lib/fetch/headers.js
+++ b/lib/fetch/headers.js
@@ -47,13 +47,16 @@ function fill (headers, object) {
       // 2. Append (header’s first item, header’s second item) to headers.
       headers.append(header[0], header[1])
     }
-  } else if (typeof object === 'object' && object !== null) {
+  } else if (object !== null && typeof object === 'object') {
     // Note: null should throw
 
     // 2. Otherwise, object is a record, then for each key → value in object,
     //    append (key, value) to headers
-    for (const [key, value] of Object.entries(object)) {
-      headers.append(key, value)
+    const keys = Object.keys(object)
+    const keysLength = keys.length
+
+    for (let i = 0; i < keysLength; i++) {
+      headers.append(keys[i], object[keys[i]])
     }
   } else {
     throw webidl.errors.conversionFailed({