@ronag 2022-11-10
@mcollina isn't there security implication to this? i.e. what happens if a partial symbol stays buffered in the encoder between requests?