@anonrig 2022-11-10
Mainly two reasons: 

1. The benchmark does not include `headers`
2. https://www.youtube.com/watch?v=jrDikOz8YSU