@KhafraDev 2022-11-10
TextEncoder's constructor only sets a single symbol, and TextDecoder seems to call into C++ land once. With those benchmarks, I don't think there's really any performance gain considering creating a new TextEncoder shouldn't be slow.